deliverables/Hack3rz_Deliverable2/Implementation/annotation-web-service-develop/deliverables/Hack3rz_Deliverable1/Implementation/annotation-web-service-develop/function-predict/local.requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
numpy==1.22.3
typing-extensions==4.1.1

# Torch for running the function locally
torch==1.10.0